requirements.txt
# Core dependencies
python-binance>=1.0.18
requests>=2.31.0
urllib3>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# CLI framework
click>=8.1.0

# Data handling and validation
pydantic>=2.0.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.16.0  # Optional: for real-time data streams

# Logging and monitoring
structlog>=23.1.0
colorama>=0.4.6  # For colored terminal output

# Mathematical operations
numpy>=1.24.0  # For advanced calculations
pandas>=2.0.0  # For data analysis (optional)

# Development and testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0  # For coverage reporting

# Code quality
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
isort>=5.12.0  # Import sorting
mypy>=1.5.0  # Type checking

# Security
cryptography>=41.0.0  # For API signature generation
certifi>=2023.7.22  # SSL certificates

# Optional: Web interface dependencies (uncomment if needed)
# fastapi>=0.100.0
# uvicorn>=0.23.0
# jinja2>=3.1.0

# Optional: Database support (uncomment if needed)
# sqlalchemy>=2.0.0
# alembic>=1.11.0

# Optional: Visualization (uncomment if needed)
# matplotlib>=3.7.0
# plotly>=5.15.0

# Optional: Machine learning for advanced strategies (uncomment if needed)
# scikit-learn>=1.3.0
# tensorflow>=2.13.0

# Pinned versions for stability (update as needed)
# Uncomment and modify if you need specific versions
# certifi==2023.7.22
# charset-normalizer==3.2.0
# idna==3.4
# python-dateutil==2.8.2
# pytz==2023.3
# six==1.16.0
# websocket-client==1.6.1